Refactor object extracton
